pyproject.toml
[tool.poetry]
name = "pantry"
version = "0.1.0"
description = ""
authors = ["Carlos G. Sarmiento <carlos@sarmiento.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["cli"], version = "^0.172.0"}
aiohttp = "^3.8.4"
aiohttp-devtools = "^1.0.post0"
tinydb = "^4.7.1"
dacite = "^1.8.0"
openfoodfacts = "^0.1.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
